+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 or argument type in luaw::function.
 * Add luavalueref::to\<T\>.
 * Add luavalueref::setglobal.
+* Rename luavalueref::getvalue to pushvalue.
 
 
 ## v1.3.0 - 2024.10.17